Move preview out + use entries
